DDLQueryBuilderFactory 생성으로 변경
