+++ src/main/java/persistence/EntityManagerImpl.java
@@ -1,23 +1,24 @@
 package persistence;
 
+import database.DatabaseServer;
 import jdbc.JdbcTemplate;
 import jdbc.PersonRowMapper;
 import persistence.entity.Person;
 import persistence.sql.ddl.DDLQueryBuilder;
+import persistence.sql.ddl.DDLQueryBuilderFactory;
 import persistence.sql.dml.DMLQueryBuilder;
 
+import java.sql.SQLException;
+
 public class EntityManagerImpl implements EntityManager {
-
-    private DDLQueryBuilder ddlQueryBuilder;
-    private DMLQueryBuilder dmlQueryBuilder;
     private final JdbcTemplate jdbcTemplate;
-
-    public EntityManagerImpl(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        ddlQueryBuilder = DDLQueryBuilder.getInstance();
-        dmlQueryBuilder = DMLQueryBuilder.getInstance();
+    private final DDLQueryBuilder ddlQueryBuilder;
+    private DMLQueryBuilder dmlQueryBuilder;
+    public EntityManagerImpl(DatabaseServer server) throws SQLException {
+        this.jdbcTemplate = new JdbcTemplate(server.getConnection());
+        this.ddlQueryBuilder = DDLQueryBuilderFactory.getDDLQueryBuilder(server);
+        this.dmlQueryBuilder = DMLQueryBuilder.getInstance();
     }
-
 
     @Override
     public <T> void createTable(Class<T> tClass) {
